requirements.txt
gradio>=4.0.0
anthropic>=0.18.0
google-genai>=0.3.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
cryptography>=42.0.0
